Excerpt bug fix
Bug: Not having text at the start of a project introduces a ‘">’ at the
top of the page.

Fix: It's expecting a piece of data called "excerpt" to be set for each
project. Without that it just grabs the first bit of HTML it sees in
the page and puts it into the <meta> description tag, which is causing
this issue.Solved by adding an `excerpt: blah blah blah` to the
metadata at the top of each project.